# services/pokedex/openapi.yml
...
openapi: 3.0.3
info:
  title: pokedex
  # Bump this when the API changes. The generated client reports it in
  # X-Client-Version, so one version covers the API and its clients.
  version: 0.2.0
  description: pokedex, owned by me-myself-and-i.
paths:
...
  /:
    get:
      operationId: getRoot
      summary: Identify the service and the running build.
      responses:
        '200':
          description: Service identity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identity'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pokemon:
    get:
      operationId: listPokemon
      summary: List Pokemon, optionally filtered by type.
      parameters:
        - name: type
          in: query
          required: false
          description: Filter to Pokemon having this type, e.g. "fire".
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Maximum number to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
      responses:
        '200':
          description: The matching Pokemon, in Pokedex order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PokemonList'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pokemon/{name}:
    get:
      operationId: getPokemon
      summary: One Pokemon by name.
      parameters:
        - name: name
          in: path
          required: true
          description: Lowercase name, e.g. "pikachu".
          schema:
            type: string
      responses:
        '200':
          description: The Pokemon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pokemon'
        '404':
          description: No Pokemon by that name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /types:
    get:
      operationId: listTypes
      summary: Every type present in the Pokedex, with a count.
      responses:
        '200':
          description: The types.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeList'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ---------------------------------------------------------------- battles
  #
  # Battles are server-authoritative: damage, turn order, legality and the
  # win condition are decided here. Clients post intents and render what
  # comes back. Three client implementations agreeing on a damage formula
  # is three chances to disagree, and a disagreement looks like cheating.
  #
  # Clients poll GET /battles/{id} and compare `version`, which increases
  # on every change. Websockets would need hand-written transport in each
  # client and an ingress that negotiates upgrades; polling is what the
  # generated client already does well.
...
components:
...
  schemas:
...
    Identity:
      type: object
      required: [service, version]
      properties:
        service:
          type: string
        version:
          type: string
    Pokemon:
      type: object
      required: [id, name, types, height, weight, sprite, moves]
      properties:
        id:
          type: integer
          description: National Pokedex number.
        name:
          type: string
        types:
          type: array
          items:
            type: string
          description: One or two types, e.g. ["grass", "poison"].
        height:
          type: integer
          description: Decimetres, as the upstream Pokedex reports it.
        weight:
          type: integer
          description: Hectograms, as the upstream Pokedex reports it.
        sprite:
          type: string
          description: Absolute URL to the front-facing sprite.
        moves:
          type: array
          items:
            $ref: '#/components/schemas/Move'
    Move:
      type: object
      required: [name, type, power]
      properties:
        name:
          type: string
        type:
          type: string
        power:
          type: integer
          description: Base power; 0 for status moves that deal no damage.
    PokemonList:
      type: object
      required: [count, pokemon]
      properties:
        count:
          type: integer
        pokemon:
          type: array
          items:
            $ref: '#/components/schemas/Pokemon'
    TypeSummary:
      type: object
      required: [name, count]
      properties:
        name:
          type: string
        count:
          type: integer
        strongAgainst:
          type: array
          description: >
            Types this one deals double damage to. Served here so the
            clients do not each carry a copy of the chart the server
            computes damage from - three copies is three chances to
            disagree with the server, and the server is what decides.
          items:
            type: string
        weakAgainst:
          type: array
          description: Types this one deals half damage to.
          items:
            type: string
        noEffectAgainst:
          type: array
          description: Types this one cannot damage at all.
          items:
            type: string
    TypeList:
      type: object
      required: [types]
      properties:
        types:
          type: array
          items:
            $ref: '#/components/schemas/TypeSummary'

    # Declaring a `default` response with this shape is what lets ogen
    # generate convenient errors: a handler can return a plain error and
    # the generated code renders it, instead of every operation needing
    # its own error union.
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string